requirements.txt
beautifulsoup4
chardet
coverage
ensae_teaching_cs
falcon>=1.4.1
folium
ijson
jyquickhelper
keras_applications
keyring
keyrings.alt
lightmlrestapi
manydataapi
mlinsights>=0.1.125
mnist
mpld3
opencv-python
pandas_streaming
papierstat
psutil
pycryptodomex
pyecharts
pyecharts-javascripthon
pyensae>=1.2.788
pymyinstall
pyquickhelper>=1.9
pyshp
qgrid
shapely
sphinx_modern_theme_modified
shapely
waitress
wheel
x86cpu
xgboost
xlrd

